+++ src/http/header/macros.rs
@@ -1,17 +1,18 @@
+// TODO: replace with derive_more impl
 macro_rules! common_header_deref {
     ($from:ty => $to:ty) => {
-        impl ::std::ops::Deref for $from {
+        impl ::core::ops::Deref for $from {
             type Target = $to;
 
             #[inline]
-            fn deref(&self) -> &$to {
+            fn deref(&self) -> &Self::Target {
                 &self.0
             }
         }
 
-        impl ::std::ops::DerefMut for $from {
+        impl ::core::ops::DerefMut for $from {
             #[inline]
-            fn deref_mut(&mut self) -> &mut $to {
+            fn deref_mut(&mut self) -> &mut Self::Target {
                 &mut self.0
             }
         }
@@ -42,14 +43,19 @@ macro_rules! common_header_test {
 
             let raw = $raw;
             let a: Vec<Vec<u8>> = raw.iter().map(|x| x.to_vec()).collect();
+
             let mut req = test::TestRequest::default();
+
             for item in a {
                 req = req.insert_header((HeaderField::name(), item)).take();
             }
+
             let req = req.finish();
             let value = HeaderField::parse(&req);
+
             let result = format!("{}", value.unwrap());
             let expected = String::from_utf8(raw[0].to_vec()).unwrap();
+
             let result_cmp: Vec<String> = result
                 .to_ascii_lowercase()
                 .split(' ')
@@ -60,9 +66,11 @@ macro_rules! common_header_test {
                 .split(' ')
                 .map(|x| x.to_owned())
                 .collect();
+
             assert_eq!(result_cmp.concat(), expected_cmp.concat());
         }
     };
+
     ($id:ident, $raw:expr, $typed:expr) => {
         #[test]
         fn $id() {
@@ -76,8 +84,10 @@ macro_rules! common_header_test {
             let req = req.finish();
             let val = HeaderField::parse(&req);
             let typed: Option<HeaderField> = $typed;
+
             // Test parsing
             assert_eq!(val.ok(), typed);
+
             // Test formatting
             if typed.is_some() {
                 let raw = &($raw)[..];
@@ -111,37 +121,41 @@ macro_rules! common_header {
             fn name() -> $crate::http::header::HeaderName {
                 $name
             }
+
             #[inline]
-            fn parse<T>(msg: &T) -> Result<Self, $crate::error::ParseError>
-                where T: $crate::HttpMessage
-            {
-                $crate::http::header::from_comma_delimited(
-                    msg.headers().get_all(Self::name())).map($id)
+            fn parse<T: $crate::HttpMessage>(msg: &T) -> Result<Self, $crate::error::ParseError> {
+                let headers = msg.headers().get_all(Self::name());
+                $crate::http::header::from_comma_delimited(headers).map($id)
             }
         }
-        impl std::fmt::Display for $id {
+
+        impl ::core::fmt::Display for $id {
             #[inline]
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> ::std::fmt::Result {
+            fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
                 $crate::http::header::fmt_comma_delimited(f, &self.0[..])
             }
         }
+
         impl $crate::http::header::IntoHeaderValue for $id {
             type Error = $crate::http::header::InvalidHeaderValue;
 
             fn try_into_value(self) -> Result<$crate::http::header::HeaderValue, Self::Error> {
-                use std::fmt::Write;
+                use ::core::fmt::Write;
                 let mut writer = $crate::http::header::Writer::new();
                 let _ = write!(&mut writer, "{}", self);
                 $crate::http::header::HeaderValue::from_maybe_shared(writer.take())
             }
         }
     };
+
     // List header, one or more items
     ($(#[$a:meta])*($id:ident, $name:expr) => ($item:ty)+) => {
         $(#[$a])*
         #[derive(Clone, Debug, PartialEq)]
         pub struct $id(pub Vec<$item>);
+
         crate::http::header::common_header_deref!($id => Vec<$item>);
+
         impl $crate::http::header::Header for $id {
             #[inline]
             fn name() -> $crate::http::header::HeaderName {
@@ -155,48 +169,54 @@ macro_rules! common_header {
                     msg.headers().get_all(Self::name())).map($id)
             }
         }
-        impl std::fmt::Display for $id {
+
+        impl ::core::fmt::Display for $id {
             #[inline]
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+            fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
                 $crate::http::header::fmt_comma_delimited(f, &self.0[..])
             }
         }
+
         impl $crate::http::header::IntoHeaderValue for $id {
             type Error = $crate::http::header::InvalidHeaderValue;
 
             fn try_into_value(self) -> Result<$crate::http::header::HeaderValue, Self::Error> {
-                use std::fmt::Write;
+                use ::core::fmt::Write;
                 let mut writer = $crate::http::header::Writer::new();
                 let _ = write!(&mut writer, "{}", self);
                 $crate::http::header::HeaderValue::from_maybe_shared(writer.take())
             }
         }
     };
+
     // Single value header
     ($(#[$a:meta])*($id:ident, $name:expr) => [$value:ty]) => {
         $(#[$a])*
         #[derive(Clone, Debug, PartialEq)]
         pub struct $id(pub $value);
+
         crate::http::header::common_header_deref!($id => $value);
+
         impl $crate::http::header::Header for $id {
             #[inline]
             fn name() -> $crate::http::header::HeaderName {
                 $name
             }
+
             #[inline]
-            fn parse<T>(msg: &T) -> Result<Self, $crate::error::ParseError>
-                where T: $crate::HttpMessage
-            {
-                $crate::http::header::from_one_raw_str(
-                    msg.headers().get(Self::name())).map($id)
+            fn parse<T: $crate::HttpMessage>(msg: &T) -> Result<Self, $crate::error::ParseError> {
+                let header = msg.headers().get(Self::name());
+                $crate::http::header::from_one_raw_str(header).map($id)
             }
         }
-        impl std::fmt::Display for $id {
+
+        impl ::core::fmt::Display for $id {
             #[inline]
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                std::fmt::Display::fmt(&self.0, f)
+            fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
+                ::core::fmt::Display::fmt(&self.0, f)
             }
         }
+
         impl $crate::http::header::IntoHeaderValue for $id {
             type Error = $crate::http::header::InvalidHeaderValue;
 
@@ -205,6 +225,7 @@ macro_rules! common_header {
             }
         }
     };
+
     // List header, one or more items with "*" option
     ($(#[$a:meta])*($id:ident, $name:expr) => {Any / ($item:ty)+}) => {
         $(#[$a])*
@@ -215,42 +236,46 @@ macro_rules! common_header {
             /// Only the listed items are a match
             Items(Vec<$item>),
         }
+
         impl $crate::http::header::Header for $id {
             #[inline]
             fn name() -> $crate::http::header::HeaderName {
                 $name
             }
-            #[inline]
-            fn parse<T>(msg: &T) -> Result<Self, $crate::error::ParseError>
-                where T: $crate::HttpMessage
-            {
-                let any = msg.headers().get(Self::name()).and_then(|hdr| {
-                    hdr.to_str().ok().and_then(|hdr| Some(hdr.trim() == "*"))});
 
-                if let Some(true) = any {
+            #[inline]
+            fn parse<T: $crate::HttpMessage>(msg: &T) -> Result<Self, $crate::error::ParseError> {
+                let is_any = msg
+                    .headers()
+                    .get(Self::name())
+                    .and_then(|hdr| hdr.to_str().ok())
+                    .map(|hdr| hdr.trim() == "*");
+
+                if let Some(true) = is_any {
                     Ok($id::Any)
                 } else {
-                    Ok($id::Items(
-                        $crate::http::header::from_comma_delimited(
-                            msg.headers().get_all(Self::name()))?))
+                    let headers = msg.headers().get_all(Self::name());
+                    Ok($id::Items($crate::http::header::from_comma_delimited(headers)?))
                 }
             }
         }
-        impl std::fmt::Display for $id {
+
+        impl ::core::fmt::Display for $id {
             #[inline]
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+            fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
                 match *self {
                     $id::Any => f.write_str("*"),
-                    $id::Items(ref fields) => $crate::http::header::fmt_comma_delimited(
-                        f, &fields[..])
+                    $id::Items(ref fields) =>
+                        $crate::http::header::fmt_comma_delimited(f, &fields[..])
                 }
             }
         }
+
         impl $crate::http::header::IntoHeaderValue for $id {
             type Error = $crate::http::header::InvalidHeaderValue;
 
             fn try_into_value(self) -> Result<$crate::http::header::HeaderValue, Self::Error> {
-                use std::fmt::Write;
+                use ::core::fmt::Write;
                 let mut writer = $crate::http::header::Writer::new();
                 let _ = write!(&mut writer, "{}", self);
                 $crate::http::header::HeaderValue::from_maybe_shared(writer.take())